change(#54): Move PageRenderer to PagesController, this seems a bit more conventional
