+++ src/Http/Controllers/PagesController.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Bambamboole\LaravelCms\Services;
+namespace Bambamboole\LaravelCms\Http\Controllers;
 
 use Bambamboole\LaravelCms\Models\Page;
 use Illuminate\Http\Request;
 
-class PageRenderer
+class PagesController
 {
-    public function render(Request $request)
+    public function show(Request $request)
     {
         $pathParts = explode('.', $request->route()->getName());
         $slug = end($pathParts);
